# Polygon in Rect performance improvements