@michaelkirk 2024-06-14
I suspect you're right that some of our predicates could be optimized for Rects, but contains is a notoriously tricky predicate.

A couple of cases that seem like they might be wrong with your implementation:

1. What about an empty polygon? By definition at least one point of P needs to be in the interior of R. That seems easy enough to guard against.

2. More tricky: What about something like P = R.to_polygon(). I'd expect your implementation to return false, because you're only checking the polygons vertices, and all the Polygon's vertices are in the boundary of R, not in the interior of R.