@michaelkirk 2024-06-15
I wonder if using `p.bounding_rect().coords_iter()` instead of `p.exterior_coords` would eke out a little more performance? Probalby not much more, but maybe a little. It also might be a good way to generalize a solution for all the other geometries too.